Refactor nvi helpers
